+++ crates/router/src/core/payments.rs
@@ -131,6 +131,10 @@ use crate::{
     core::{
         errors::{self, CustomResult, RouterResponse, RouterResult},
         payment_methods::{cards, network_tokenization},
+        payments::helpers::{
+            get_applepay_metadata, is_applepay_predecrypted_flow_supported,
+            is_googlepay_predecrypted_flow_supported,
+        },
         payouts,
         routing::{self as core_routing},
         unified_authentication_service::types::{ClickToPay, UnifiedAuthenticationService},
@@ -3966,13 +3970,6 @@ where
             _ => return Ok(None),
         };
 
-        // Check if the wallet has already decrypted the token from the payment data.
-        // If a pre-decrypted token is available, use it directly to avoid redundant decryption.
-        if let Some(predecrypted_token) = wallet.check_predecrypted_token(payment_data)? {
-            logger::debug!("Using predecrypted token for wallet");
-            return Ok(Some(predecrypted_token));
-        }
-
         let merchant_connector_account =
             get_merchant_connector_account_for_wallet_decryption_flow::<F, D>(
                 state,
@@ -3981,6 +3978,15 @@ where
                 connector_call_type_optional,
             )
             .await?;
+        // Check if the wallet has already decrypted the token from the payment data and support_predecrypted_token is true.
+        // If a pre-decrypted token is available, check if it is enabled in mca and use it directly to avoid redundant decryption.
+
+        if let Some(predecrypted_token) =
+            wallet.check_predecrypted_token(payment_data, &merchant_connector_account)?
+        {
+            logger::debug!("Using predecrypted token for wallet");
+            return Ok(Some(predecrypted_token));
+        }
 
         let decide_wallet_flow = &wallet
             .decide_wallet_flow(state, payment_data, &merchant_connector_account)
@@ -4419,6 +4425,7 @@ where
     ) {
         let payment_method_data = payment_data.get_payment_method_data();
         let customer_id = payment_data.get_payment_intent().customer_id.clone();
+
         if let (Some(domain::PaymentMethodData::Card(card_data)), Some(customer_id)) =
             (payment_method_data, customer_id)
         {
@@ -5802,6 +5809,7 @@ where
     fn check_predecrypted_token(
         &self,
         _payment_data: &D,
+        _merchant_connector_account: &helpers::MerchantConnectorAccountType,
     ) -> CustomResult<Option<PaymentMethodToken>, errors::ApiErrorResponse> {
         // Default implementation returns None (no pre-decrypted data)
         Ok(None)
@@ -5898,15 +5906,16 @@ where
     fn check_predecrypted_token(
         &self,
         payment_data: &D,
+        merchant_connector_account: &helpers::MerchantConnectorAccountType,
     ) -> CustomResult<Option<PaymentMethodToken>, errors::ApiErrorResponse> {
         let apple_pay_wallet_data = payment_data
             .get_payment_method_data()
             .and_then(|payment_method_data| payment_method_data.get_wallet_data())
             .and_then(|wallet_data| wallet_data.get_apple_pay_wallet_data());
 
-        let result = if let Some(data) = apple_pay_wallet_data {
+        if let Some(data) = apple_pay_wallet_data {
             match &data.payment_data {
-                common_payments_types::ApplePayPaymentData::Encrypted(_) => None,
+                common_payments_types::ApplePayPaymentData::Encrypted(_) => Ok(None),
                 common_payments_types::ApplePayPaymentData::Decrypted(
                     apple_pay_predecrypt_data,
                 ) => {
@@ -5914,15 +5923,24 @@ where
                         &apple_pay_predecrypt_data.application_expiration_month,
                         &apple_pay_predecrypt_data.application_expiration_year,
                     )?;
-                    Some(PaymentMethodToken::ApplePayDecrypt(Box::new(
-                        apple_pay_predecrypt_data.clone(),
-                    )))
+
+                    if is_applepay_predecrypted_flow_supported(
+                        merchant_connector_account.get_metadata(),
+                    ) {
+                        Ok(Some(PaymentMethodToken::ApplePayDecrypt(Box::new(
+                            apple_pay_predecrypt_data.clone(),
+                        ))))
+                    } else {
+                        Err(errors::ApiErrorResponse::PreconditionFailed {
+                            message: "Predecrypted token is not enabled for ApplePay.".to_string(),
+                        }
+                        .into())
+                    }
                 }
             }
         } else {
-            None
-        };
-        Ok(result)
+            Ok(None)
+        }
     }
 
     fn decide_wallet_flow(
@@ -5940,10 +5958,10 @@ where
         );
 
         let wallet_flow = match apple_pay_metadata {
-            Some(domain::ApplePayFlow::Simplified(payment_processing_details)) => Some(
+            Some(domain::ApplePayFlow::DecryptAtApplication(payment_processing_details)) => Some(
                 DecideWalletFlow::ApplePayDecrypt(payment_processing_details),
             ),
-            Some(domain::ApplePayFlow::Manual) | None => None,
+            Some(domain::ApplePayFlow::SkipDecryption) | None => None,
         };
         Ok(wallet_flow)
     }
@@ -6010,31 +6028,41 @@ where
     fn check_predecrypted_token(
         &self,
         payment_data: &D,
+        merchant_connector_account: &helpers::MerchantConnectorAccountType,
     ) -> CustomResult<Option<PaymentMethodToken>, errors::ApiErrorResponse> {
         let google_pay_wallet_data = payment_data
             .get_payment_method_data()
             .and_then(|payment_method_data| payment_method_data.get_wallet_data())
             .and_then(|wallet_data| wallet_data.get_google_pay_wallet_data());
 
-        let result = if let Some(data) = google_pay_wallet_data {
+        if let Some(data) = google_pay_wallet_data {
             match &data.tokenization_data {
-                common_payments_types::GpayTokenizationData::Encrypted(_) => None,
+                common_payments_types::GpayTokenizationData::Encrypted(_) => Ok(None),
                 common_payments_types::GpayTokenizationData::Decrypted(
                     google_pay_predecrypt_data,
                 ) => {
-                    helpers::validate_card_expiry(
-                        &google_pay_predecrypt_data.card_exp_month,
-                        &google_pay_predecrypt_data.card_exp_year,
-                    )?;
-                    Some(PaymentMethodToken::GooglePayDecrypt(Box::new(
-                        google_pay_predecrypt_data.clone(),
-                    )))
+                    if is_googlepay_predecrypted_flow_supported(
+                        merchant_connector_account.get_metadata(),
+                    ) {
+                        helpers::validate_card_expiry(
+                            &google_pay_predecrypt_data.card_exp_month,
+                            &google_pay_predecrypt_data.card_exp_year,
+                        )?;
+                        Ok(Some(PaymentMethodToken::GooglePayDecrypt(Box::new(
+                            google_pay_predecrypt_data.clone(),
+                        ))))
+                    } else {
+                        Err(errors::ApiErrorResponse::PreconditionFailed {
+                            message: "Predecrypted token flow is not enabled for GooglePay."
+                                .to_string(),
+                        }
+                        .into())
+                    }
                 }
             }
         } else {
-            None
-        };
-        Ok(result)
+            Ok(None)
+        }
     }
     fn decide_wallet_flow(
         &self,
@@ -7501,73 +7529,62 @@ fn check_apple_pay_metadata(
     merchant_connector_account.and_then(|mca| {
         let metadata = mca.get_metadata();
         metadata.and_then(|apple_pay_metadata| {
-            let parsed_metadata = apple_pay_metadata
-                .clone()
-                .parse_value::<api_models::payments::ApplepayCombinedSessionTokenData>(
-                    "ApplepayCombinedSessionTokenData",
-                )
-                .map(|combined_metadata| {
-                    api_models::payments::ApplepaySessionTokenMetadata::ApplePayCombined(
-                        combined_metadata.apple_pay_combined,
-                    )
-                })
-                .or_else(|_| {
-                    apple_pay_metadata
-                        .parse_value::<api_models::payments::ApplepaySessionTokenData>(
-                            "ApplepaySessionTokenData",
-                        )
-                        .map(|old_metadata| {
-                            api_models::payments::ApplepaySessionTokenMetadata::ApplePay(
-                                old_metadata.apple_pay,
-                            )
-                        })
-                })
-                .map_err(|error| {
-                    logger::warn!(?error, "Failed to Parse Value to ApplepaySessionTokenData")
-                });
+            let parsed_metadata = get_applepay_metadata(Some(apple_pay_metadata.clone()));
 
-            parsed_metadata.ok().map(|metadata| match metadata {
+            parsed_metadata.ok().and_then(|metadata| match metadata {
                 api_models::payments::ApplepaySessionTokenMetadata::ApplePayCombined(
                     apple_pay_combined,
-                ) => match apple_pay_combined {
-                    api_models::payments::ApplePayCombinedMetadata::Simplified { .. } => {
-                        domain::ApplePayFlow::Simplified(payments_api::PaymentProcessingDetails {
-                            payment_processing_certificate: state
-                                .conf
-                                .applepay_decrypt_keys
-                                .get_inner()
-                                .apple_pay_ppc
-                                .clone(),
-                            payment_processing_certificate_key: state
-                                .conf
-                                .applepay_decrypt_keys
-                                .get_inner()
-                                .apple_pay_ppc_key
-                                .clone(),
-                        })
+                ) => match apple_pay_combined.get_combined_metadata_required() {
+                    Ok(api_models::payments::ApplePayCombinedMetadata::Simplified { .. }) => {
+                        Some(domain::ApplePayFlow::DecryptAtApplication(
+                            payments_api::PaymentProcessingDetails {
+                                payment_processing_certificate: state
+                                    .conf
+                                    .applepay_decrypt_keys
+                                    .get_inner()
+                                    .apple_pay_ppc
+                                    .clone(),
+                                payment_processing_certificate_key: state
+                                    .conf
+                                    .applepay_decrypt_keys
+                                    .get_inner()
+                                    .apple_pay_ppc_key
+                                    .clone(),
+                            },
+                        ))
                     }
-                    api_models::payments::ApplePayCombinedMetadata::Manual {
+                    Ok(api_models::payments::ApplePayCombinedMetadata::Manual {
                         payment_request_data: _,
                         session_token_data,
-                    } => {
+                    }) => {
                         if let Some(manual_payment_processing_details_at) =
                             session_token_data.payment_processing_details_at
                         {
                             match manual_payment_processing_details_at {
                                 payments_api::PaymentProcessingDetailsAt::Hyperswitch(
                                     payment_processing_details,
-                                ) => domain::ApplePayFlow::Simplified(payment_processing_details),
+                                ) => Some(domain::ApplePayFlow::DecryptAtApplication(
+                                    payment_processing_details,
+                                )),
                                 payments_api::PaymentProcessingDetailsAt::Connector => {
-                                    domain::ApplePayFlow::Manual
+                                    Some(domain::ApplePayFlow::SkipDecryption)
                                 }
                             }
                         } else {
-                            domain::ApplePayFlow::Manual
+                            Some(domain::ApplePayFlow::SkipDecryption)
+                        }
+                    }
+                    Err(_) => {
+                        // In the case were only predecrypted token in enabled donot throw error , just skip decryption
+                        if apple_pay_combined.is_predecrypted_token_supported() {
+                            Some(domain::ApplePayFlow::SkipDecryption)
+                        } else {
+                            None
                         }
                     }
                 },
                 api_models::payments::ApplepaySessionTokenMetadata::ApplePay(_) => {
-                    domain::ApplePayFlow::Manual
+                    Some(domain::ApplePayFlow::SkipDecryption)
                 }
             })
         })
@@ -7594,12 +7611,10 @@ fn get_google_pay_connector_wallet_details(
                 });
 
             google_pay_wallet_details
-                .ok()
+                .ok().map(|details| details.google_pay)
                 .and_then(
                     |google_pay_wallet_details| {
-                        match google_pay_wallet_details
-                        .google_pay
-                        .provider_details {
+                        match google_pay_wallet_details.provider_details {
                             api_models::payments::GooglePayProviderDetails::GooglePayMerchantDetails(merchant_details) => {
                                 match (
                                     merchant_details